Lession_5/seminar_test/requirements.txt
numpy
pandas
matplotlib
emoji
python-telegram-bot
